Refactor build chain
