+++ package.json
@@ -16,6 +16,7 @@
   },
   "author": "NDLA (https://ndla.no)",
   "license": "MIT",
+  "type": "module",
   "devDependencies": {
     "@axe-core/react": "^4.7.3",
     "@babel/cli": "^7.23.0",
@@ -53,5 +54,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-tiny-popover": "^7.2.4"
-  }
+  },
+  "browserslist": "> 0.5%, last 3 versions, Firefox ESR, not dead"
 }
