+++ src/entries/dist.js
@@ -1,0 +1,4 @@
+import OrderPriority from '@scripts/app.js';
+
+// Load library
+H5P.OrderPriority = OrderPriority;
