+++ src/scripts/components/Surface/components/StatementList/components/components/StatementComment.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import { useOrderPriority } from 'context/OrderPriorityContext';
+import { useOrderPriority } from '@context/OrderPriorityContext.js';
 
 /**
  * Comment displayed "inline", not in a popover
