+++ src/scripts/context/OrderPriorityContext.js
@@ -2,21 +2,25 @@ import React, { useContext } from 'react';
 
 const OrderPriorityContext = React.createContext();
 
-function OrderPriorityProvider({ children, value }) {
+const OrderPriorityProvider = ({ children, value }) => {
   return (
     <OrderPriorityContext.Provider value={value}>
       {children}
     </OrderPriorityContext.Provider>
   );
-}
+};
 
-function useOrderPriority() {
+const useOrderPriority = () => {
   const context = useContext(OrderPriorityContext);
+
   if ( context === undefined) {
-    throw new Error('useOrderPriority must be used within a OrderPriorityProvider');
+    throw new Error(
+      'useOrderPriority must be used within a OrderPriorityProvider'
+    );
   }
+
   return context;
-}
+};
 
 export {
   OrderPriorityProvider,
