+++ src/styles/components/Placeholder.scss
